fix multi-thread problem
