+++ saga/seata-saga-engine-store/src/main/java/io/seata/saga/engine/store/db/DbAndReportTcStateLogStore.java
@@ -396,9 +396,11 @@ public class DbAndReportTcStateLogStore extends AbstractStore implements StateLo
         if (isUpdateMode) {
             return originalCompensateStateInstId + "-" + maxIndex;
         }
-        for (StateInstance aStateInstance : stateInstance.getStateMachineInstance().getStateList()) {
-            if (aStateInstance != stateInstance
-                    && originalCompensateStateInstId.equals(aStateInstance.getStateIdCompensatedFor())) {
+
+        for (int i = 0; i < stateInstance.getStateMachineInstance().getStateList().size(); i++) {
+            StateInstance aStateInstance = stateInstance.getStateMachineInstance().getStateList().get(i);
+            if (aStateInstance != stateInstance && originalCompensateStateInstId.equals(
+                aStateInstance.getStateIdCompensatedFor())) {
                 int idIndex = getIdIndex(aStateInstance.getId(), "-");
                 maxIndex = idIndex > maxIndex ? idIndex : maxIndex;
                 maxIndex++;
@@ -440,8 +442,8 @@ public class DbAndReportTcStateLogStore extends AbstractStore implements StateLo
 
         } else if (StringUtils.hasLength(stateInstance.getStateIdCompensatedFor())) {
 
-            // find if this compensate has been executed
-            for (StateInstance aStateInstance : stateInstance.getStateMachineInstance().getStateList()) {
+            for (int i = 0; i < stateInstance.getStateMachineInstance().getStateList().size(); i++) {
+                StateInstance aStateInstance = stateInstance.getStateMachineInstance().getStateList().get(i);
                 if (aStateInstance.isForCompensation() && aStateInstance.getName().equals(stateInstance.getName())) {
                     if (null != state.isCompensatePersistModeUpdate()) {
                         return state.isCompensatePersistModeUpdate();
