+++ saga/seata-saga-engine/src/main/java/io/seata/saga/engine/pcext/utils/LoopContextHolder.java
@@ -57,8 +57,6 @@ public class LoopContextHolder {
 
     public static void clearCurrent(ProcessContext context) {
         context.removeVariable(DomainConstants.VAR_NAME_CURRENT_LOOP_CONTEXT_HOLDER);
-        context.removeVariable(DomainConstants.VAR_NAME_IS_LOOP_STATE);
-        context.removeVariable(DomainConstants.VAR_NAME_CURRENT_LOOP_STATE);
     }
 
     public AtomicInteger getNrOfInstances() {
