+++ test/src/test/java/io/seata/saga/engine/db/mockserver/StateMachineDBMockServerTests.java
@@ -30,9 +30,7 @@ import org.junit.jupiter.api.Test;
 import org.springframework.context.ApplicationContext;
 import org.springframework.context.support.ClassPathXmlApplicationContext;
 
-import java.util.ArrayList;
 import java.util.HashMap;
-import java.util.List;
 import java.util.Map;
 
 /**
@@ -673,106 +671,5 @@ public class StateMachineDBMockServerTests {
         StateMachineInstance instance = stateMachineEngine.getStateMachineConfig().getStateLogStore().getStateMachineInstance(
                 "10.15.232.93:8091:2019567124");
         System.out.println(instance);
-    }
-
-    @Test
-    public void testSimpleStateMachineWithLoopForward() throws InterruptedException {
-        long start  = System.currentTimeMillis();
-
-        List<Integer> loopList = new ArrayList<>();
-        for (int i = 0; i < 10; i++) {
-            loopList.add(i);
-        }
-
-        Map<String, Object> paramMap = new HashMap<>(2);
-        paramMap.put("a", 1);
-        paramMap.put("collection", loopList);
-        paramMap.put("fooThrowException", "true");
-
-        String stateMachineName = "simpleLoopTestStateMachine";
-
-        StateMachineInstance inst = stateMachineEngine.start(stateMachineName, null, paramMap);
-
-        long cost = System.currentTimeMillis() - start;
-        System.out.println("====== cost :" + cost);
-
-        Assertions.assertEquals(inst.getStatus(), ExecutionStatus.UN);
-
-        Thread.sleep(1000);
-
-        inst = stateMachineEngine.forward(inst.getId(), paramMap);
-        Assertions.assertEquals(inst.getStatus(), ExecutionStatus.UN);
-
-        paramMap.put("fooThrowException", "false");
-        inst = stateMachineEngine.forward(inst.getId(), paramMap);
-        Assertions.assertEquals(inst.getStatus(), ExecutionStatus.SU);
-    }
-
-    @Test
-    public void testSimpleStateMachineWithLoopCompensateForRecovery() throws InterruptedException {
-        long start  = System.currentTimeMillis();
-
-        List<Integer> loopList = new ArrayList<>();
-        for (int i = 0; i < 10; i++) {
-            loopList.add(i);
-        }
-
-        Map<String, Object> paramMap = new HashMap<>(2);
-        paramMap.put("a", 1);
-        paramMap.put("collection", loopList);
-        paramMap.put("barThrowException", "true");
-        paramMap.put("compensateFooThrowException", "true");
-
-        String stateMachineName = "simpleLoopTestStateMachine";
-
-        StateMachineInstance inst = stateMachineEngine.start(stateMachineName, null, paramMap);
-
-        long cost = System.currentTimeMillis() - start;
-        System.out.println("====== cost :" + cost);
-
-        Assertions.assertEquals(inst.getStatus(), ExecutionStatus.UN);
-        Assertions.assertEquals(inst.getCompensationStatus(), ExecutionStatus.UN);
-
-        Thread.sleep(1000);
-
-        inst = stateMachineEngine.compensate(inst.getId(), paramMap);
-        Assertions.assertEquals(inst.getCompensationStatus(), ExecutionStatus.UN);
-
-        paramMap.put("compensateFooThrowException", "false");
-        inst = stateMachineEngine.compensate(inst.getId(), paramMap);
-        Assertions.assertEquals(inst.getCompensationStatus(), ExecutionStatus.SU);
-    }
-
-    @Test
-    public void testSimpleStateMachineWithLoopSubMachineForward() throws InterruptedException {
-        long start = System.currentTimeMillis();
-
-        List<Integer> loopList = new ArrayList<>();
-        for (int i = 0; i < 10; i++) {
-            loopList.add(i);
-        }
-
-        Map<String, Object> paramMap = new HashMap<>(2);
-        paramMap.put("a", 2);
-        paramMap.put("collection", loopList);
-        paramMap.put("barThrowException", "true");
-
-        String stateMachineName = "simpleLoopTestStateMachine";
-
-        StateMachineInstance inst = stateMachineEngine.start(stateMachineName, null, paramMap);
-
-        long cost = System.currentTimeMillis() - start;
-        System.out.println("====== cost :" + cost);
-
-        Assertions.assertEquals(inst.getStatus(), ExecutionStatus.UN);
-
-        Thread.sleep(1000);
-
-        inst = stateMachineEngine.forward(inst.getId(), paramMap);
-        Assertions.assertEquals(inst.getStatus(), ExecutionStatus.UN);
-
-        paramMap.put("barThrowException", "false");
-        inst = stateMachineEngine.forward(inst.getId(), paramMap);
-        Assertions.assertEquals(inst.getStatus(), ExecutionStatus.SU);
     }
 }
