+++ test/src/test/resources/saga/spring/statemachine_engine_test.xml
@@ -35,7 +35,7 @@
 	<bean id="threadExecutor"
 		  class="org.springframework.scheduling.concurrent.ThreadPoolExecutorFactoryBean">
 		<property name="threadNamePrefix" value="SAGA_ASYNC_EXE_" />
-		<property name="corePoolSize" value="20" />
+		<property name="corePoolSize" value="1" />
 		<property name="maxPoolSize" value="20" />
 		<property name="queueCapacity" value="100" />
 		<property name="rejectedExecutionHandler" ref="callerRunsPolicy" />
